hw3 chore: moved tests
