+++ currency-rates/tests/load_tests/load.go
